move each command to its own module
